Move the subscription types to the eth-rpc
